refactor: moved NPM related GraphQL extensions to self-contained module
